+++ src/DependencyInjection/Configuration.php
@@ -52,6 +52,7 @@ class Configuration implements ConfigurationInterface
     {
         $subscriptions = $root->arrayNode('subscriptions')
                               ->info('Manage subscriptions settings')
+                              ->canBeDisabled()
                               ->addDefaultsIfNotSet()
                               ->children();